// src/core/enum/error.enum.ts
export enum ErrorMessage {
  ACCOUNT_NOT_EXISTS = 'error.ACCOUNT_NOT_EXISTS',
  ACCOUNT_EXISTS = 'error.ACCOUNT_EXISTS',
  PASSWORD_INCORRECT = 'error.PASSWORD_INCORRECT',
  REFRESH_TOKEN_NOT_EXISTS = 'error.REFRESH_TOKEN_NOT_EXISTS',
  UNAUTHORIZED = 'error.UNAUTHORIZED',
  INVALID_PARAM = 'error.INVALID_PARAM',
  SYSTEM_REGISTER_CODE_INCORRECT = 'error.SYSTEM_REGISTER_CODE_INCORRECT',
  RESTAURANT_NOT_EXISTS = 'error.RESTAURANT_NOT_EXISTS',
  MENU_CATEGORY_NOT_EXISTS = 'error.MENU_CATEGORY_NOT_EXISTS',
}